Added threshold in the counting of nulls

+++ lunaticEngine/src/it/unibas/lunatic/model/chase/chasede/operators/ChaseDEScenarioProxy.java
@@ -102,7 +102,7 @@ public class ChaseDEScenarioProxy implements IDEChaser {
                 System.out.println("# " + tableName + ": " + tableSize + " tuples");
             }
         }
-        long numberOfNulls = databaseAnalyzer.countNulls(targetDB);
+        Integer numberOfNulls = databaseAnalyzer.countNulls(targetDB);
         System.out.println("# Number of nulls: " + numberOfNulls);
         System.out.println("### Total Number of Tuples: " + totalNumberOfTuples + " tuples");
         long preProcessingTime = 0L;
